chat now aligns properly on page load without having to refresh
