+++ app/javascript/plugins/init_chat_alignments.js
@@ -1,0 +1,17 @@
+const initChatAlignments = () => {
+
+  const currentUserId = $(".chat-container").first().attr('data-my-user-id');
+  const currentUserMessages = document.querySelectorAll("div[data-message-userid=\""+currentUserId+"\"]");
+
+  currentUserMessages.forEach(message => {
+    moveChatLogRight(message);
+  })
+
+  function moveChatLogRight(message) {
+    message.classList.add("my-messages");
+  }
+
+
+};
+
+export { initChatAlignments };
